+++ Makefile.am
@@ -129,7 +129,7 @@ src_libfabric_la_SOURCES =			\
 	include/ofi_iov.h			\
 	include/fi_list.h			\
 	include/ofi_lock.h			\
-	include/fi_mem.h			\
+	include/ofi_mem.h			\
 	include/fi_osd.h			\
 	include/ofi_proto.h			\
 	include/ofi_rbuf.h			\
