+++ prov/netdir/src/netdir_iface.h
@@ -41,7 +41,7 @@
 
 #include "rdma/fabric.h"
 
-#include "fi_mem.h"
+#include "ofi_mem.h"
 #include "fi_list.h"
 #include "rdma/fi_eq.h"
 #include "rdma/fi_domain.h"
